+++ scripts/full-lean/probes/host.cpp
@@ -4,9 +4,9 @@
 #include <cstdlib>
 #include <pthread.h>
 
-extern "C" int32_t lasm_node_call(uint32_t, uint32_t, uint64_t, const uint8_t *, uint32_t);
-extern "C" void lasm_node_copy(uint8_t *, uint32_t);
-extern "C" uint32_t lasm_node_start(uint32_t, uint32_t, uint64_t, const uint8_t *, uint32_t);
+extern "C" int64_t lasm_node_call(uint32_t, uint32_t, uint64_t, const uint8_t *, size_t);
+extern "C" void lasm_node_copy(uint8_t *, size_t);
+extern "C" uint32_t lasm_node_start(uint32_t, uint32_t, uint64_t, const uint8_t *, size_t);
 
 static void *run(void *) {
     // Sleep resolves asynchronously on the host while this pthread blocks.
